+++ prisma/migrations/20220224124910_alter_image_types_to_uri/migration.sql
@@ -1,0 +1,14 @@
+/*
+  Warnings:
+
+  - You are about to drop the column `image` on the `User` table. All the data in the column will be lost.
+  - Added the required column `logoUri` to the `Team` table without a default value. This is not possible if the table is not empty.
+  - Added the required column `avatarUri` to the `User` table without a default value. This is not possible if the table is not empty.
+
+*/
+-- AlterTable
+ALTER TABLE "Team" ADD COLUMN     "logoUri" TEXT NOT NULL;
+
+-- AlterTable
+ALTER TABLE "User" DROP COLUMN "image",
+ADD COLUMN     "avatarUri" TEXT NOT NULL;