+++ prisma/migrations/20220224125408_alter_image_uris_to_nullable/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "Team" ALTER COLUMN "logoUri" DROP NOT NULL;
+
+-- AlterTable
+ALTER TABLE "User" ALTER COLUMN "avatarUri" DROP NOT NULL;