+++ src/users/users.module.ts
@@ -2,9 +2,10 @@ import { Module } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
 import { TeamsService } from 'src/teams/teams.service';
+import { StorageService } from 'src/storage/storage.service';
 
 @Module({
   controllers: [UsersController],
-  providers: [UsersService, TeamsService],
+  providers: [UsersService, TeamsService, StorageService],
 })
 export class UsersModule {}